client: refactor code and allow custom http clients
