+++ eywa.go
@@ -4,53 +4,8 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"io"
-	"net/http"
 	"strings"
 )
-
-type Client struct {
-	endpoint   string
-	httpClient *http.Client
-	headers    map[string]string
-}
-
-func NewClient(endpoint string, headers map[string]string) *Client {
-	return &Client{
-		endpoint:   endpoint,
-		httpClient: http.DefaultClient,
-		headers:    headers,
-	}
-}
-
-func (c *Client) do(q string) (*bytes.Buffer, error) {
-	reqObj := graphqlRequest{
-		Query: q,
-	}
-
-	var reqBytes bytes.Buffer
-	err := json.NewEncoder(&reqBytes).Encode(&reqObj)
-	if err != nil {
-		return nil, err
-	}
-	req, err := http.NewRequest(http.MethodPost, c.endpoint, &reqBytes)
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Add("Content-Type", "application/json")
-	for key, value := range c.headers {
-		req.Header.Add(key, value)
-	}
-	resp, err := c.httpClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	var respBytes bytes.Buffer
-	_, err = io.Copy(&respBytes, resp.Body)
-	return &respBytes, err
-}
 
 type graphqlRequest struct {
 	Query     string                 `json:"query"`
